reduced number of lib files
