--- src/lib/object-filter.js
@@ -1,5 +1,0 @@
-export const objectFilter = (filter = {}) => (obj = {}) =>
-  Object.keys(filter).reduce(
-    (out, key) => out && filter[key] === obj[key],
-    true
-  )
